+++ src/commands/action.rb
@@ -16,19 +16,3 @@ class Action
     Bot.say_as_poke(sender, msg)
   end
 end
-
-class SimpleAction < Action
-  def initialize
-    super { run } 
-  end
-end
-
-class WithArgumentsAction < Action
-  def initialize
-    super { |with_message| run(with_message) }
-  end
-
-  def run(with_message)
-    raise "Not implemented yet."
-  end
-end
