+++ src/commands/simple_action.rb
@@ -1,0 +1,5 @@
+class SimpleAction < Action
+  def initialize
+    super { run } 
+  end
+end
